1 byte (uint8_t) is not big enough to store potential set element text str node numbers (as used by CMEX) since these are regular size int (4 byte)

+++ src/gxfile.cpp
@@ -2997,7 +2997,7 @@ int TGXFileObj::gdxSetTextNodeNr( int TxtNr, int Node )
    auto &obj = *SetTextList;
    if( TxtNr >= 0 && TxtNr < obj.size() && !*obj.GetObject( TxtNr ) )
    {
-      *obj.GetObject( TxtNr ) = static_cast<uint8_t>( Node );
+      *obj.GetObject( TxtNr ) = Node;
       return true;
    }
    return false;
